backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Image Processing
exifread==3.0.0
piexif==1.1.3
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scipy==1.11.4
colour-science==0.4.3

# Utilities
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
